[AI] Update to vertigo 4.3

+++ src/main/java/io/vertigo/ai/impl/bb/BlackBoardImpl.java
@@ -180,7 +180,7 @@ final class BlackBoardImpl implements BlackBoard {
 	//- All methods are prefixed with list
 	//------------------------------------
 	@Override
-	public int listSize(final BBKey key) {
+	public long listSize(final BBKey key) {
 		checkKey(key, Type.List);
 		//---
 		return blackBoardStorePlugin
